+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Next Release
 * GoBook 1-B-DupLines
 * GoBook 1-C-Lissajous
 * GoBook 1-D-FetchURL
+* GoBook 1-E-WebServer
 
 Version 1.4.0
 -